server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
python-dotenv
google-cloud-storage
Werkzeug
